Separate gun firing into triggered and firing
Triggered is set by the act of inputting the fire button or an AI trying to shoot and read by the actual firing sytem.
Firing is set by the actual firing system and read by display and audio systems.

This allows for things like jamming guns, visibly rotating rotary cannons even when they're out of ammo, that sorta thing.